backend/requirements.txt
torch
transformers
datasets
scikit-learn
peft
flask
flask-cors
flask_caching
psycopg2-binary
